streaming-app/spark-streaming/requirements.txt
kafka-python==2.0.2
python-dotenv==1.0.1
selenium==4.27.1
PyYAML==6.0.1
requests==2.31.0
flask==2.3.3
prometheus_client==0.20.0
pymongo==4.7.0